% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis.R
\name{campsismiscOutfun}
\alias{campsismiscOutfun}
\title{Campsismisc output function.}
\usage{
campsismiscOutfun(x, outputs)
}
\arguments{
\item{x}{tibble}

\item{outputs}{oat_analysis_outputs}
}
\value{
tibble
}
\description{
Campsismisc output function.
}
