% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_covariate.R
\name{CategoricalLabeledCovariate}
\alias{CategoricalLabeledCovariate}
\title{Create a categorical labeled covariate.}
\usage{
CategoricalLabeledCovariate(
  name,
  default_value,
  label = NULL,
  unit = NULL,
  categories
)
}
\arguments{
\item{name}{covariate name, single character value}

\item{default_value}{default covariate value, single numeric value}

\item{label}{covariate label, e.g. 'Weight'}

\item{unit}{value unit, e.g. 'kg'}

\item{categories}{named vector describing the categories, e.g. c(FED=0, FASTED=1)}
}
\value{
a labeled covariate
}
\description{
Create a categorical labeled covariate.
}
