% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_change.R
\name{Change}
\alias{Change}
\title{Create a sensibility analysis change.}
\usage{
Change(parameter, up, down, upDownAsFactor = TRUE)
}
\arguments{
\item{parameter}{parameter name, like "CL", must be part of the model}

\item{up}{model parameter will be multiplied by this value if upDownAsFactor is TRUE,
otherwise the model parameter will take this value as is}

\item{down}{model parameter will be divided by this value if upDownAsFactor is TRUE,
otherwise the model parameter will take this value as is}

\item{upDownAsFactor}{if TRUE, up and down are respectively used as multiplier and divider,
if FALSE, their value is used as is}
}
\description{
Create a sensibility analysis change.
}
