% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_factor.R
\name{Factor}
\alias{Factor}
\title{Create a sensibility analysis factor.}
\usage{
Factor(parameter, multiplier)
}
\arguments{
\item{parameter}{parameter name, like "CL", must be part of the model}

\item{multiplier}{multiplication factor to be analysed}
}
\description{
Create a sensibility analysis factor.
}
