% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_output.R
\docType{class}
\name{forest_plot_output-class}
\alias{forest_plot_output-class}
\title{Forest plot output, see this class as an interface.}
\description{
Forest plot output, see this class as an interface.
}
