% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{ForestPlot}
\alias{ForestPlot}
\title{Create a forest plot.}
\usage{
ForestPlot(
  model,
  outputs,
  dataset = NULL,
  replicates = 1L,
  dest = "mrgsolve",
  settings = Settings()
)
}
\arguments{
\item{model}{CAMPSIS model}

\item{outputs}{output(s) of interest, e.g. 'CL' or a PK metric from campsisnca or a list of them}

\item{dataset}{CAMPSIS dataset, if NULL, minimalist dataset with 1 subject 
and single observation at time 0 is created}

\item{replicates}{number of replicates}

\item{dest}{destination engine: rxode2, RxODE or mrgsolve (default)}

\item{settings}{simulation settings}
}
\value{
an empty forest plot
}
\description{
Create a forest plot.
}
