% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_item.R
\name{ForestPlotItem}
\alias{ForestPlotItem}
\title{Create a forest plot item.}
\usage{
ForestPlotItem(...)
}
\arguments{
\item{...}{all constant covariates}
}
\value{
a forest plot item
}
\description{
Create a forest plot item.
}
