% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_items.R
\name{ForestPlotItems}
\alias{ForestPlotItems}
\title{Create a list of forest plot items}
\usage{
ForestPlotItems()
}
\value{
a list of forest plot items
}
\description{
Create a list of forest plot items
}
