% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/oat_analysis.R
\name{getForestPlot}
\alias{getForestPlot}
\alias{getForestPlot,oat_analysis,integer,logical,logical,logical,logical,numeric,numeric-method}
\title{Get forest plot.}
\usage{
getForestPlot(
  object,
  index = NULL,
  relative = NULL,
  show_labels = NULL,
  show_ref = NULL,
  show_range = NULL,
  range = NULL,
  ci = NULL,
  ...
)

\S4method{getForestPlot}{oat_analysis,integer,logical,logical,logical,logical,numeric,numeric}(
  object,
  index,
  relative,
  show_labels,
  show_ref,
  show_range,
  range,
  ci,
  geom_label_vjust = 0,
  geom_label_nudge_x = 0.15,
  geom_label_nudge_y = 0,
  geom_label_size = 3,
  label_nsig = 3,
  geom_hline_color = "darkblue",
  ...
)
}
\arguments{
\item{object}{generic object}

\item{index}{output number, 1 by default}

\item{relative}{relative change or absolute value, logical}

\item{show_labels}{show the numeric labels}

\item{show_ref}{show vertical reference (at 1 or at the baseline)}

\item{show_range}{show vertical dashed lines, logical}

\item{range}{vertical dashed lines range, default is c(0.8, 1.25)}

\item{ci}{confidence interval, default is 0.9, which corresponds to a 90\% CI}

\item{...}{extra arguments like 'geom_label_vjust', 'geom_label_nudge_x', 'geom_label_nudge_y', 'geom_label_size', 'label_nsig', 'geom_hline_color'}
}
\value{
ggplot
}
\description{
Get forest plot.
}
