% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis_output.R
\name{getOutvars}
\alias{getOutvars}
\alias{getOutvars,model_parameter_output-method}
\alias{getOutvars,nca_metric_output-method}
\title{Get out variables to be simulated.}
\usage{
getOutvars(object)

\S4method{getOutvars}{model_parameter_output}(object)

\S4method{getOutvars}{nca_metric_output}(object)
}
\arguments{
\item{object}{output}
}
\value{
variables to be simulated
}
\description{
Get out variables to be simulated.
}
