% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/forest_plot.R
\name{getPlot}
\alias{getPlot}
\alias{getPlot,forest_plot-method}
\title{Get plot.}
\usage{
getPlot(object, ...)

\S4method{getPlot}{forest_plot}(object)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
ggplot
}
\description{
Get plot.
}
