% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/oat_analysis.R
\name{getTornadoPlot}
\alias{getTornadoPlot}
\alias{getTornadoPlot,oat_analysis,integer,logical,logical,logical-method}
\title{Get tornado plot.}
\usage{
getTornadoPlot(
  object,
  index = NULL,
  relative = NULL,
  show_labels = NULL,
  show_ref = NULL,
  ...
)

\S4method{getTornadoPlot}{oat_analysis,integer,logical,logical,logical}(
  object,
  index,
  relative,
  show_labels,
  show_ref,
  geom_bar_width = 0.5,
  geom_text_nudge_y = 1,
  label_nsig = 3,
  geom_hline_color = "grey",
  geom_text_size = 3,
  ...
)
}
\arguments{
\item{object}{generic object}

\item{index}{output number, 1 by default}

\item{relative}{relative change or absolute value, logical}

\item{show_labels}{show the numeric labels}

\item{show_ref}{show vertical reference (at 0 or at the baseline)}

\item{...}{extra arguments like 'geom_bar_width', 'geom_text_nudge_y', 'label_nsig', 'geom_hline_color', 'geom_text_size'}

\item{geom_bar_width}{width argument of method geom_bar}

\item{geom_text_nudge_y}{nudge_y argument of method geom_text}

\item{label_nsig}{number of significant digits in label}

\item{geom_hline_color}{color argument of method geom_hline}

\item{geom_text_size}{size argument of method geom_text}
}
\value{
ggplot
}
\description{
Get tornado plot.
}
