% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_covariate.R
\name{LabeledCovariate}
\alias{LabeledCovariate}
\title{Create a labeled covariate.}
\usage{
LabeledCovariate(name, value, label = NULL, value_label = NULL, unit = NULL)
}
\arguments{
\item{name}{covariate name, single character value}

\item{value}{covariate value, single numeric value to use in CAMPSIS model}

\item{label}{covariate label, e.g. 'Weight'}

\item{value_label}{value label (useful for discrete covariate), e.g. 'Fasted'}

\item{unit}{value unit, e.g. 'kg'}
}
\value{
a labeled covariate
}
\description{
Create a labeled covariate.
}
