% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_covariates.R
\name{LabeledCovariates}
\alias{LabeledCovariates}
\title{Create a list of labeled covariates.}
\usage{
LabeledCovariates()
}
\value{
a list of labeled covariates
}
\description{
Create a list of labeled covariates.
}
