% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_parameters.R
\name{LabeledParameters}
\alias{LabeledParameters}
\title{Add labels to your parameters.}
\usage{
LabeledParameters(x)
}
\arguments{
\item{x}{named vector of parameters (names are the parameter names, values are the corresponding labels)}
}
\value{
a labeled_parameters object
}
\description{
Add labels to your parameters.
}
