% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis_output.R
\name{ModelParameterOutput}
\alias{ModelParameterOutput}
\title{Create a model parameter output.}
\usage{
ModelParameterOutput(name)
}
\arguments{
\item{name}{model parameter name}
}
\description{
Create a model parameter output.
}
