% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis_output.R
\docType{class}
\name{oat_analysis_output-class}
\alias{oat_analysis_output-class}
\title{OAT analysis output, see this class as an interface.}
\description{
OAT analysis output, see this class as an interface.
}
