% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis_result.R
\name{OATResult}
\alias{OATResult}
\title{Create OAT result structure.}
\usage{
OATResult(output, baseline, results)
}
\arguments{
\item{output}{corresponding output}

\item{baseline}{baseline value}

\item{results}{post-processed results}
}
\value{
OAT result structure
}
\description{
Create OAT result structure.
}
