% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/oat_analysis.R
\name{prepare}
\alias{prepare}
\alias{prepare,oat_analysis-method}
\title{Prepare.}
\usage{
prepare(object)

\S4method{prepare}{oat_analysis}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
same object
}
\description{
Prepare.
}
