% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_changes.R
\docType{class}
\name{sensitivity_analysis_changes-class}
\alias{sensitivity_analysis_changes-class}
\title{Sensitivity analysis changes class.}
\description{
Sensitivity analysis changes class.
}
