% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_factors.R
\docType{class}
\name{sensitivity_analysis_factors-class}
\alias{sensitivity_analysis_factors-class}
\title{Sensitivity analysis factors class.}
\description{
Sensitivity analysis factors class.
}
