% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_item.R
\docType{class}
\name{sensitivity_analysis_item-class}
\alias{sensitivity_analysis_item-class}
\title{Forest plot item: one or more labeled covariates.}
\description{
Forest plot item: one or more labeled covariates.
}
