% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{SensitivityAnalysis}
\alias{SensitivityAnalysis}
\title{Create an OAT-based sensitivity analysis object.}
\usage{
SensitivityAnalysis(
  model,
  output,
  dataset = NULL,
  replicates = 1L,
  dest = "mrgsolve",
  settings = Settings()
)
}
\arguments{
\item{model}{CAMPSIS model}

\item{output}{output of interest, e.g. 'CL' or a PK metric from campsisnca}

\item{dataset}{CAMPSIS dataset, if NULL, minimalist dataset with 1 subject 
and single observation at time 0 is created}

\item{replicates}{number of replicates}

\item{dest}{destination engine: rxode2, RxODE or mrgsolve (default)}

\item{settings}{simulation settings}
}
\value{
an empty OAT-based sensibility analysis object
}
\description{
Create an OAT-based sensitivity analysis object.
}
