% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_changes.R
\name{SensitivityAnalysisChanges}
\alias{SensitivityAnalysisChanges}
\title{Create a list of sensitivity analysis changes}
\usage{
SensitivityAnalysisChanges()
}
\value{
a list of sensitivity analysis changes
}
\description{
Create a list of sensitivity analysis changes
}
