% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_factors.R
\name{SensitivityAnalysisFactors}
\alias{SensitivityAnalysisFactors}
\title{Create a list of sensitivity analysis factors}
\usage{
SensitivityAnalysisFactors()
}
\value{
a list of sensitivity analysis factors
}
\description{
Create a list of sensitivity analysis factors
}
