% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_item.R
\name{SensitivityAnalysisItem}
\alias{SensitivityAnalysisItem}
\title{Create a sensitivity analysis item.}
\usage{
SensitivityAnalysisItem(...)
}
\arguments{
\item{...}{all factors}
}
\value{
a sensitivity analysis item
}
\description{
Create a sensitivity analysis item.
}
