% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_items.R
\name{SensitivityAnalysisItems}
\alias{SensitivityAnalysisItems}
\title{Create a list of sensitivity analysis items}
\usage{
SensitivityAnalysisItems()
}
\value{
a list of sensitivity analysis items
}
\description{
Create a list of sensitivity analysis items
}
