% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis.R, R/forest_plot.R
\name{createScenarios}
\alias{createScenarios}
\alias{createScenarios,forest_plot-method}
\title{Create scenarios.}
\usage{
createScenarios(object, ...)

\S4method{createScenarios}{forest_plot}(object, dataset)
}
\arguments{
\item{object}{OAT analysis, type forest plot or sensitivity analysis}

\item{...}{extra arguments}
}
\value{
a list of CAMPSIS scenarios to be run
}
\description{
Create scenarios.
}
