% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{ForestPlot}
\alias{ForestPlot}
\title{Create a forest plot.}
\usage{
ForestPlot(model, output, dataset = NULL, replicates = 1L)
}
\arguments{
\item{model}{CAMPSIS model}

\item{output}{output of interest, e.g. 'CL' or a PK metric from campsisnca}

\item{dataset}{CAMPSIS dataset, if NULL, minimalist dataset with 1 subject 
and single observation at time 0 is created}

\item{replicates}{number of replicates}
}
\value{
an empty forest plot
}
\description{
Create a forest plot.
}
