% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/oat_analysis.R, R/forest_plot.R
\name{getForestPlot}
\alias{getForestPlot}
\alias{getForestPlot,oat_analysis,logical,logical,logical,logical,numeric,numeric,numeric,numeric-method}
\alias{getForestPlot,forest_plot,logical,logical,logical,logical,numeric,numeric,numeric,numeric-method}
\title{Get forest plot.}
\usage{
getForestPlot(
  object,
  relative = NULL,
  show_labels = NULL,
  show_ref = NULL,
  show_range = NULL,
  range = NULL,
  ci = NULL,
  limits = NULL,
  breaks = NULL,
  ...
)

\S4method{getForestPlot}{oat_analysis,logical,logical,logical,logical,numeric,numeric,numeric,numeric}(
  object,
  relative = NULL,
  show_labels = NULL,
  show_ref = NULL,
  show_range = NULL,
  range = NULL,
  ci = NULL,
  limits = NULL,
  breaks = NULL,
  ...
)

\S4method{getForestPlot}{forest_plot,logical,logical,logical,logical,numeric,numeric,numeric,numeric}(
  object,
  relative = NULL,
  show_labels = NULL,
  show_ref = NULL,
  show_range = NULL,
  range = NULL,
  ci = NULL,
  limits = NULL,
  breaks = NULL,
  ...
)
}
\arguments{
\item{object}{generic object}

\item{relative}{relative change or absolute value, logical}

\item{show_labels}{show the numeric labels}

\item{show_ref}{show vertical reference (at 1 or at the baseline)}

\item{show_range}{show vertical dashed lines, logical}

\item{range}{vertical dashed lines range, default is c(0.8, 1.25)}

\item{ci}{confidence interval, default is 0.9, which corresponds to a 90\% CI}

\item{limits}{ggplot limits}

\item{breaks}{ggplot breaks}

\item{...}{extra arguments}
}
\value{
ggplot
}
\description{
Get forest plot.
}
