% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/forest_plot_item.R,
%   R/sensitivity_analysis_item.R
\name{getLabel}
\alias{getLabel}
\alias{getLabel,forest_plot_item-method}
\alias{getLabel,sensitivity_analysis_item-method}
\title{Get label.}
\usage{
getLabel(object, ...)

\S4method{getLabel}{forest_plot_item}(object, labeled_covariates)

\S4method{getLabel}{sensitivity_analysis_item}(object, labeled_parameters)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
a label
}
\description{
Get label.
}
