% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/forest_plot.R
\name{getPlot}
\alias{getPlot}
\alias{getPlot,forest_plot-method}
\title{Get plot.}
\usage{
getPlot(object, ...)

\S4method{getPlot}{forest_plot}(
  object,
  limits = c(0.5, 1.5),
  breaks = c(0.7, 0.8, 1, 1.25, 1.4),
  vjust = 0,
  nudge_x = 0.15,
  nudge_y = 0,
  size = 3
)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
ggplot
}
\description{
Get plot.
}
