% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/oat_analysis.R
\name{getTornadoPlot}
\alias{getTornadoPlot}
\alias{getTornadoPlot,sensitivity_analysis,logical,logical,logical-method}
\title{Get tornado plot.}
\usage{
getTornadoPlot(
  object,
  relative = NULL,
  show_labels = NULL,
  show_ref = NULL,
  ...
)

\S4method{getTornadoPlot}{sensitivity_analysis,logical,logical,logical}(
  object,
  relative,
  show_labels,
  show_ref,
  geom_bar_color = "#94c0e3",
  geom_bar_width = 0.5,
  geom_text_nudge_y = 1,
  ...
)
}
\arguments{
\item{object}{generic object}

\item{relative}{relative change or absolute value, logical}

\item{show_labels}{show the numeric labels}

\item{show_ref}{show vertical reference (at 0 or at the baseline)}

\item{...}{extra arguments like 'geom_bar_color', 'geom_bar_width', 'geom_text_nudge_y'}
}
\value{
ggplot
}
\description{
Get tornado plot.
}
