% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat_analysis_output.R
\name{NcaMetricOutput}
\alias{NcaMetricOutput}
\title{Create a model parameter output.}
\usage{
NcaMetricOutput(metric)
}
\arguments{
\item{name}{model parameter name}
}
\description{
Create a model parameter output.
}
