% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/forest_plot.R
\name{prepare}
\alias{prepare}
\alias{prepare,forest_plot-method}
\title{Prepare.}
\usage{
prepare(object)

\S4method{prepare}{forest_plot}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
same object
}
\description{
Prepare.
}
